Refactor code and move to TPVConfigDumper
